requirements.txt
boto3==1.17.83
botocore==1.20.83
certifi==2020.4.5.1
chardet==3.0.4
cpools==0.0.0
cycler==0.10.0
decorator==4.4.2
et-xmlfile==1.0.1
ffmpeg-python==0.2.0
future==0.18.2
h5py==2.10.0
idna==2.9
imageio==2.9.0
jdcal==1.4.1
jmespath==0.10.0
joblib==1.0.1
jsonpatch==1.25
jsonpointer==2.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.19.5
olefile==0.46
opencv-python==4.1.2.30
pandas==1.1.5
Pillow==7.0.0
protobuf==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
rank-bm25==0.2.1
regex==2021.4.4
requests==2.23.0
s3transfer==0.4.2
scikit-image==0.16.2
scikit-learn==0.24.2
scipy==1.4.1
six==1.16.0
sklearn==0.0
tensorboardX==2.0
threadpoolctl==2.1.0
torch==1.4.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.5.0
tornado==6.0.4
tqdm==4.43.0
urllib3==1.25.9
visdom==0.1.8.9
websocket-client==0.57.0
